Move agent.Tracer -> worker.tracer (unexported)
